pull EditField out of Modal to be able to prototype better
